Tried a simplet state system, not what I want
Lowering the number of states comes at the cost of increased complexity for varying the number of descriptions of scenes, as well as tracking the availability of scenes. I.e. if there's a mirror scene, the scene previous should not have affordance for it if the mirror has been taken. Ultimately a PoC that lead to the realization of a bad idea.